+++ src/xterm.c
@@ -1289,8 +1289,8 @@ x_draw_glyphless_glyph_string_foreground (struct glyph_string *s)
       else if (glyph->u.glyphless.method == GLYPHLESS_DISPLAY_HEX_CODE)
 	{
 	  snprintf(buf, sizeof(buf), "%0*X",
-		   ((glyph->u.glyphless.ch < 0x10000) ? 4 : 6),
-		   (unsigned int)glyph->u.glyphless.ch);
+		   (uint8_t)((glyph->u.glyphless.ch < 0x10000) ? 4 : 6),
+		   (unsigned short)glyph->u.glyphless.ch);
 	  str = buf;
 	}
